Update service_config doc
The current grpclb fallback implementation is to continue using the internal round_robin policy but feed it with the backends from the resolver. We didn't fall back to the policy indicated in service config (or client API) because that's too complex and has more different behavior from the normal situation where we do connect to the balancer.