Return UNKNOWN on GenericRpcHandler failure
A GenericRpcHandler registered on a gRPC Server is not supposed to raise
an exception and if it does so it is considered a programming defect.
However, gRPC is supposed to respond to the client with an UNKNOWN
status code.  Previously, this situation was left unhandled and the
client ended up receiving a response with CANCELLED status code.

This commit fixes the issue https://github.com/grpc/grpc/issues/13629.