Make logging after success in jobset more apparent
tools/run_tests/artifacts/artifact_targets.py is used in the artifact
build process to create the relevant jobset.JobSpec, but currently
prints the console output only if the build fails.

jobset.JobSpec has a verbose_success flag that is used to print the
process output even if it ends in SUCCESS, but is currently not easily
perceivable and is hard to find, making it very difficult to debug
SUCCESS cases. Make the flag more apparent by adding it as parameters
to jobspec creation methods in artifact_targets.