Implement a lock-free fast path for queue_call_request()
For tiny RPCs, every single requests in almost the first item in the
list. Hence, it would try to lock the server to process pending
requests.

Instead of locking, simply set and check atomic values when there is a
possiblity of having pending requests.

This increases QPS by 10%, for the 62-channel/0B-RPC benchmark using the
callback API.