Fix error in usage of _SC_THREAD_STACK_MIN in test
_SC_THREAD_STACK_MIN is intended to be used as an argument to sysconf
and not by itself.  This may have slipped past testing on x86 due to
the round up to page size, but failed on aarch64 because aarch64 has
conservative stack guard requirements to account for different page
sizes (4K and 64K).

Sysconfs get used correctly elsewhere in this file, so this is likely
just a typo.

Signed-off-by: Siddhesh Poyarekar <siddhesh@linaro.org>