Fixing ListenerRemoved test.
In the ListenerRemoved test, we observed that XdsConfigSelector gets re-created after being destroyed upon removal of listener.

This sometimes causes test to fail as RPC will continue to succeed which is not expected; more importantly the re-creation of XdsConfigSelector is not the correct behaviour.

XdsConfigSelector gets recreated because the XdsConfigSelector destructor calls will check on removed clusters and then
recreate XdsConfigSelector due to the change.

This recreation uses a stored copy of the LDS/RDS update, the correct solution is to clear that update upon the removal of listener (ResourceDoesNotExist) and thus prvent the recreation of XdsConfigSelector.

Fix is tested to make sure ListenerRemoved test is no longer flaky; logs are checked to ensure that after Listener Removal, the CDs and EDs requests contain no resources and the XDS server will unsubscribe to those resources and no RPC will succeed.