Add darwin_arm64 and darwin_arm64e to c-ares' config settings
This fixes Bazel selecting the wrong source for `cares` when building
with `--cpu=darwin_arm64` or `--cpu=darwin_arm64e` (it was selecting the
source in the default condition, which is for Linux).

We're having the same patch in Bazel
(https://github.com/bazelbuild/bazel/commit/1782f0ae751569607ef88930c822ac460a1f8bb3;
it was using an older grpc release). Fixing this here so we can remove
the patch in the next update.